implement handlers for the rest of the AuthorizationState messages
Created a new abstract class, BaseInput, which is meant for asking the
user questions, which is necessary when you first start the app since
`tdlib/tdjson` needs to log in. This is abstract because in case someday
this app has a user interface, we can switch out implementations

right now there is just TTYInput which asks the user questions on the
command line/TTY

the new methods in handlers.py TdlibBaseMessageHandler implement the
known possible authentication states, will ask the user questions
if necessary and then send the result to tdlib. There are a few
that are meant for 'we are logging out' (aka you go to a different
telegram app -> active sessions and terminate the session we are using)
that don't do anything at the moment but log a message.

but as far as i know, all of this works, I don't have another phone
number to test the 'registration' flow (authorizationStateWaitRegistration)
but it seems simple enough that it should work fine as is